refactor: reorganize entries and topics page folders for clarity
Rename dynamic parameter folders to semantic action names since routing is
memory-based, not file-path-based. Flattens the folder structure and improves
discoverability.

- new/ → create/
- [id]/ → detail/
- edit/[id]/ → edit/
