+++ src/entrypoints/sidepanel/pages/library/entries/create/entry-create.tsx
