+++ src/entrypoints/sidepanel/pages/library/entries/detail/entry-detail.tsx
