+++ src/entrypoints/sidepanel/pages/library/entries/edit/entry-edit.tsx
