+++ src/entrypoints/sidepanel/pages/library/topics/detail/topic-detail.tsx
